requirements.txt
Flask==2.2.2
flask-sqlalchemy==3.0.2
line-bot-sdk==2.3.0
transitions==0.9.0
graphviz==0.20.1
pygraphviz==1.10.0
yapf
pyright
toml
